+++ gql/lib/link.dart
@@ -3,6 +3,7 @@
 library link;
 
 import "package:gql/execution.dart";
+import "package:gql/language.dart";
 import "package:meta/meta.dart";
 
 /// Type of the `forward` function
@@ -45,4 +46,54 @@ class _LinkChain implements Link {
         forward,
         (fw, link) => (op) => link.request(op, fw),
       )(request);
+}
+
+/// JSON [Response] parser
+class ResponseParser {
+  /// Parses the response body
+  ///
+  /// Extend this to add non-standard behavior
+  Response parseResponse(Map<String, dynamic> body) => Response(
+        errors: (body["errors"] as List)
+            ?.map(
+              (dynamic error) => parseError(error as Map<String, dynamic>),
+            )
+            ?.toList(),
+        data: body["data"] as Map<String, dynamic>,
+      );
+
+  /// Parses a response error
+  ///
+  /// Extend this to add non-standard behavior
+  GraphQLError parseError(Map<String, dynamic> error) => GraphQLError(
+        message: error["message"] as String,
+        path: error["path"] as List,
+        locations: (error["locations"] as List)
+            ?.map(
+              (dynamic error) => parseLocation(error as Map<String, dynamic>),
+            )
+            ?.toList(),
+        extensions: error["extensions"] as Map<String, dynamic>,
+      );
+
+  /// Parses a response error location
+  ///
+  /// Extend this to add non-standard behavior
+  ErrorLocation parseLocation(Map<String, dynamic> location) => ErrorLocation(
+        line: location["line"] as int,
+        column: location["column"] as int,
+      );
+}
+
+/// JSON [Request] serializer.
+class RequestSerializer {
+  /// Serializes the request
+  ///
+  /// Extend this to add non-standard behavior
+  @visibleForOverriding
+  Map<String, dynamic> serializeRequest(Request request) => <String, dynamic>{
+        "operationName": request.operation.operationName,
+        "variables": request.operation.variables,
+        "query": printNode(request.operation.document),
+      };
 }