+++ gql/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.8.1
+
+- `package:gql/link`: implement `ResponseParser` and `RequestSerializer`
+
 ## 0.8.0
 
 - `package:gql/execution`: add `Context` to `Request`